refactor Vue declaration into shim file - [CORE-849]
